pyproject.toml
[tool.poetry]
name = "backendsignature-app"
version = "0.1.0"
description = ""
authors = ["VuTrungNguyen <vu.nguyen71@live.vu.edu.au>"]

[tool.poetry.dependencies]
python = "^3.10"
jwt = "^1.3.1"
Flask = "^2.3.3"
mysqlclient = "^2.2.0"
mysql-connector-python = "^8.1.0"
cryptography = "^41.0.3"
Flask-SQLAlchemy = "^3.0.5"
Flask-MySQLdb = "^1.0.1"

[tool.poetry.dev-dependencies]
jwt = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
